refactor: edge crate, quic_listener
