+++ crates/edge/src/quic_listener/mod.rs
@@ -60,49 +60,23 @@ use crate::{
     watchdog::{WatchdogCoordinator, WatchdogRuntimeConfig, now_millis},
 };
 
-/// A leaky token-bucket rate limiter for new QUIC connection accepts.
-///
-/// Tokens refill at `rate_per_sec` tokens/second up to a cap of `burst`.
-/// Each new `quiche::accept` call consumes one token; if the bucket is empty
-/// the packet is silently dropped (no panic, no connection state allocated).
-pub(crate) struct TokenBucket {
-    /// Maximum tokens the bucket can hold (burst capacity).
-    burst: f64,
-    /// Tokens added per nanosecond (= rate_per_sec / 1_000_000_000).
-    tokens_per_ns: f64,
-    /// Current available tokens.
-    tokens: f64,
-    /// Last time tokens were refilled.
-    last_refill: Instant,
-}
+mod connection;
+mod control_api;
+mod forwarding;
+mod health_check;
+mod token_bucket;
+mod validation;
 
-impl TokenBucket {
-    fn new(rate_per_sec: u32, burst: u32) -> Self {
-        let burst = (burst.max(1)) as f64;
-        let rate_per_sec = rate_per_sec.max(1) as f64;
-        Self {
-            burst,
-            tokens_per_ns: rate_per_sec / 1_000_000_000.0,
-            tokens: burst, // start full so the first burst of legitimate connections succeeds
-            last_refill: Instant::now(),
-        }
-    }
-
-    /// Try to consume one token. Returns `true` if a token was available
-    /// (connection may proceed), `false` if the bucket is empty (drop).
-    fn try_consume(&mut self) -> bool {
-        let now = Instant::now();
-        let elapsed_ns = now.saturating_duration_since(self.last_refill).as_nanos() as f64;
-        self.last_refill = now;
-        self.tokens = (self.tokens + elapsed_ns * self.tokens_per_ns).min(self.burst);
-        if self.tokens >= 1.0 {
-            self.tokens -= 1.0;
-            true
-        } else {
-            false
-        }
-    }
-}
+pub(crate) use connection::{ConnectionRoutes, purge_connection_routes, sweep_closed_connections};
+pub(crate) use token_bucket::TokenBucket;
+#[cfg(test)]
+use health_check::classify_active_health_check_response;
+use connection::resolve_primary_from_radix_prefix;
+use forwarding::{ForwardRequestMeta, abort_stream};
+use validation::{
+    RequestBufferError, extract_header_value, generated_span_id, generated_trace_id,
+    parse_traceparent, request_content_length, validate_request_headers,
+};
 
 fn is_hop_header(name: &str) -> bool {
     matches!(
@@ -132,463 +106,6 @@ type ResolvedBackend = (
     bool,
     RouteDecisionReason,
 );
-
-struct RequestValidationResult {
-    method: String,
-    path: String,
-    authority: Option<String>,
-}
-
-#[derive(Clone)]
-struct ForwardRequestMeta {
-    method: Arc<str>,
-    path: Arc<str>,
-    authority: Option<Arc<str>>,
-    headers: Arc<Vec<quiche::h3::Header>>,
-    client_addr: SocketAddr,
-    request_id: u64,
-    traceparent: Option<Arc<str>>,
-}
-
-impl ForwardRequestMeta {
-    fn build_bodyless_request(
-        &self,
-        endpoint: &BackendEndpoint,
-    ) -> Result<Request<BoxBody<Bytes, std::convert::Infallible>>, ProxyError> {
-        build_h2_request_for_endpoint(
-            endpoint,
-            &self.method,
-            &self.path,
-            self.headers.as_slice(),
-            BoxBody::new(Full::new(Bytes::new())),
-            Some(0),
-            ForwardedContext {
-                client_addr: self.client_addr,
-                request_authority: self.authority.as_deref(),
-                request_id: self.request_id,
-                traceparent: self.traceparent.as_deref(),
-            },
-        )
-        .map_err(ProxyError::from)
-    }
-}
-
-#[derive(Clone)]
-struct ControlApiPaths {
-    health_path: String,
-    ready_path: String,
-    runtime_path: String,
-    restart_path: String,
-}
-
-#[derive(Clone)]
-struct ControlApiState {
-    metrics: Arc<Metrics>,
-    resilience: Arc<RuntimeResilience>,
-    watchdog: Arc<WatchdogCoordinator>,
-    upstream_pools: HashMap<String, Arc<RwLock<UpstreamPool>>>,
-    expected_workers: usize,
-    started_at: Instant,
-    auth_token: Option<String>,
-}
-
-impl ControlApiState {
-    fn snapshot_backend_health(&self) -> (usize, usize) {
-        let mut healthy = 0usize;
-        let mut total = 0usize;
-        for pool in self.upstream_pools.values() {
-            let guard = match pool.read() {
-                Ok(guard) => guard,
-                Err(_) => continue,
-            };
-            let pool_total = guard.pool.len();
-            total = total.saturating_add(pool_total);
-            healthy = healthy.saturating_add(guard.pool.healthy_len().min(pool_total));
-        }
-        (healthy, total)
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RequestBufferError {
-    StreamCap,
-    GlobalCap,
-}
-
-fn request_content_length(headers: &[quiche::h3::Header]) -> Option<usize> {
-    for header in headers {
-        if !header.name().eq_ignore_ascii_case(b"content-length") {
-            continue;
-        }
-        let value = std::str::from_utf8(header.value()).ok()?;
-        let parsed = value.trim().parse::<usize>().ok()?;
-        return Some(parsed);
-    }
-    None
-}
-
-fn validate_request_headers(
-    list: &[quiche::h3::Header],
-    resilience: &RuntimeResilience,
-) -> Result<RequestValidationResult, (http::StatusCode, &'static [u8], bool)> {
-    if list.len() > resilience.max_headers_count {
-        return Err((
-            http::StatusCode::REQUEST_HEADER_FIELDS_TOO_LARGE,
-            b"too many request headers\n",
-            false,
-        ));
-    }
-
-    let mut header_bytes = 0usize;
-    let mut method = None::<String>;
-    let mut path = None::<String>;
-    let mut authority = None::<String>;
-    let mut host = None::<String>;
-    let mut scheme_seen = false;
-
-    for header in list {
-        header_bytes = header_bytes.saturating_add(header.name().len() + header.value().len());
-        if header_bytes > resilience.max_headers_bytes {
-            return Err((
-                http::StatusCode::REQUEST_HEADER_FIELDS_TOO_LARGE,
-                b"request headers exceed size limit\n",
-                false,
-            ));
-        }
-
-        match header.name() {
-            b":method" => {
-                if method.is_some() {
-                    return Err((
-                        http::StatusCode::BAD_REQUEST,
-                        b"duplicate :method header\n",
-                        false,
-                    ));
-                }
-                method = Some(String::from_utf8_lossy(header.value()).to_string());
-            }
-            b":path" => {
-                if path.is_some() {
-                    return Err((
-                        http::StatusCode::BAD_REQUEST,
-                        b"duplicate :path header\n",
-                        false,
-                    ));
-                }
-                path = Some(String::from_utf8_lossy(header.value()).to_string());
-            }
-            b":authority" => {
-                if authority.is_some() {
-                    return Err((
-                        http::StatusCode::BAD_REQUEST,
-                        b"duplicate :authority header\n",
-                        false,
-                    ));
-                }
-                authority = Some(String::from_utf8_lossy(header.value()).to_string());
-            }
-            b"host" => {
-                if host.is_some() {
-                    return Err((
-                        http::StatusCode::BAD_REQUEST,
-                        b"duplicate host header\n",
-                        false,
-                    ));
-                }
-                host = Some(String::from_utf8_lossy(header.value()).to_string());
-            }
-            b":scheme" => {
-                if scheme_seen {
-                    return Err((
-                        http::StatusCode::BAD_REQUEST,
-                        b"duplicate :scheme header\n",
-                        false,
-                    ));
-                }
-                scheme_seen = true;
-            }
-            name if name.starts_with(b":") => {
-                return Err((
-                    http::StatusCode::BAD_REQUEST,
-                    b"unsupported pseudo-header\n",
-                    false,
-                ));
-            }
-            _ => {}
-        }
-    }
-
-    let method = match method {
-        Some(method) => method,
-        None => {
-            return Err((
-                http::StatusCode::BAD_REQUEST,
-                b"missing :method header\n",
-                false,
-            ));
-        }
-    };
-    let path = match path {
-        Some(path) => path,
-        None => {
-            return Err((
-                http::StatusCode::BAD_REQUEST,
-                b"missing :path header\n",
-                false,
-            ));
-        }
-    };
-
-    if method.trim().is_empty() || method.as_bytes().iter().any(|b| b.is_ascii_whitespace()) {
-        return Err((
-            http::StatusCode::BAD_REQUEST,
-            b"invalid :method header\n",
-            false,
-        ));
-    }
-
-    if path.is_empty() || !path.starts_with('/') {
-        return Err((
-            http::StatusCode::BAD_REQUEST,
-            b"invalid :path header\n",
-            false,
-        ));
-    }
-
-    if resilience.enforce_authority_host_match
-        && let (Some(authority_value), Some(host_value)) = (authority.as_deref(), host.as_deref())
-    {
-        let normalized_authority = normalize_host_for_routing(authority_value)
-            .unwrap_or_else(|| authority_value.to_ascii_lowercase());
-        let normalized_host = normalize_host_for_routing(host_value)
-            .unwrap_or_else(|| host_value.to_ascii_lowercase());
-        if normalized_authority != normalized_host {
-            return Err((
-                http::StatusCode::BAD_REQUEST,
-                b":authority and host headers must match\n",
-                false,
-            ));
-        }
-    }
-
-    if !resilience.method_allowed(&method) {
-        return Err((
-            http::StatusCode::METHOD_NOT_ALLOWED,
-            b"request method blocked by policy\n",
-            true,
-        ));
-    }
-
-    if resilience.path_denied(&path) {
-        return Err((
-            http::StatusCode::FORBIDDEN,
-            b"request path blocked by policy\n",
-            true,
-        ));
-    }
-
-    Ok(RequestValidationResult {
-        method,
-        path,
-        authority: authority.or(host),
-    })
-}
-
-fn extract_header_value<'a>(headers: &'a [quiche::h3::Header], name: &[u8]) -> Option<&'a str> {
-    headers
-        .iter()
-        .find(|header| header.name().eq_ignore_ascii_case(name))
-        .and_then(|header| std::str::from_utf8(header.value()).ok())
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-}
-
-fn parse_traceparent(value: &str) -> Option<(String, String)> {
-    let parts: Vec<&str> = value.split('-').collect();
-    if parts.len() != 4 {
-        return None;
-    }
-    if parts[0] != "00" {
-        return None;
-    }
-    let trace_id = parts[1];
-    let parent_span_id = parts[2];
-    let flags = parts[3];
-
-    let trace_valid = trace_id.len() == 32
-        && trace_id.chars().all(|c| c.is_ascii_hexdigit())
-        && trace_id != "00000000000000000000000000000000";
-    let span_valid = parent_span_id.len() == 16
-        && parent_span_id.chars().all(|c| c.is_ascii_hexdigit())
-        && parent_span_id != "0000000000000000";
-    let flags_valid = flags.len() == 2 && flags.chars().all(|c| c.is_ascii_hexdigit());
-
-    if !(trace_valid && span_valid && flags_valid) {
-        return None;
-    }
-
-    Some((
-        trace_id.to_ascii_lowercase(),
-        parent_span_id.to_ascii_lowercase(),
-    ))
-}
-
-fn generated_trace_id(conn_trace_id: &str, request_id: u64) -> String {
-    let mut seed = conn_trace_id.as_bytes().to_vec();
-    seed.extend_from_slice(&request_id.to_be_bytes());
-    let lo = crate::stable_hash64(&seed);
-    seed.extend_from_slice(b"trace-hi");
-    let hi = crate::stable_hash64(&seed);
-    format!("{hi:016x}{lo:016x}")
-}
-
-fn generated_span_id(request_id: u64) -> String {
-    format!("{:016x}", crate::stable_hash64(&request_id.to_be_bytes()))
-}
-
-fn resolve_primary_from_radix_prefix<T>(
-    dcid: &[u8],
-    connections: &HashMap<Arc<[u8]>, T>,
-    cid_routes: &mut HashMap<Arc<[u8]>, Arc<[u8]>>,
-    cid_radix: &mut CidRadix,
-) -> Option<Arc<[u8]>> {
-    let matched_cid = cid_radix.longest_prefix_match(dcid)?;
-
-    if connections.contains_key(matched_cid.as_ref()) {
-        return Some(matched_cid);
-    }
-
-    if let Some(primary) = cid_routes.get(matched_cid.as_ref()).cloned() {
-        if connections.contains_key(primary.as_ref()) {
-            return Some(primary);
-        }
-        // Alias points to a missing primary, clean up stale alias map entry.
-        cid_routes.remove(matched_cid.as_ref());
-    }
-
-    // Stale radix entry (either alias or retired primary).
-    cid_radix.remove(matched_cid.as_ref());
-    None
-}
-
-/// Remove all routing index entries for a connection that is being torn down.
-///
-/// Cleans `cid_routes`, `cid_radix`, and `peer_routes` for both the primary
-/// SCID and every alias SCID in `routing_scids`.  Extracted as a free
-/// function so it can be exercised in unit tests without a full
-/// `QUICListener` instance.
-pub(crate) fn purge_connection_routes(
-    cid_routes: &mut HashMap<Arc<[u8]>, Arc<[u8]>>,
-    cid_radix: &mut CidRadix,
-    peer_routes: &mut HashMap<std::net::SocketAddr, Arc<[u8]>>,
-    primary_scid: &Arc<[u8]>,
-    routing_scids: &std::collections::HashSet<Arc<[u8]>>,
-    peer_address: &std::net::SocketAddr,
-) {
-    cid_radix.remove(primary_scid.as_ref());
-    cid_routes.remove(primary_scid.as_ref());
-    for cid in routing_scids {
-        cid_radix.remove(cid.as_ref());
-        cid_routes.remove(cid.as_ref());
-    }
-    peer_routes.remove(peer_address);
-}
-
-/// Routing metadata extracted from a connection for use during cleanup.
-/// Kept as a plain struct so `sweep_closed_connections` is testable without
-/// a real `quiche::Connection`.
-pub(crate) struct ConnectionRoutes {
-    pub primary_scid: Arc<[u8]>,
-    pub routing_scids: std::collections::HashSet<Arc<[u8]>>,
-    pub peer_address: std::net::SocketAddr,
-}
-
-impl From<&crate::QuicConnection> for ConnectionRoutes {
-    fn from(c: &crate::QuicConnection) -> Self {
-        Self {
-            primary_scid: Arc::clone(&c.primary_scid),
-            routing_scids: c.routing_scids.clone(),
-            peer_address: c.peer_address,
-        }
-    }
-}
-
-/// Remove every SCID in `to_remove` from `connections` and all routing
-/// indexes.  Called from `handle_timeouts` after the closed-connection scan
-/// and extracted here so the removal sweep can be tested without a live
-/// `QUICListener`.
-pub(crate) fn sweep_closed_connections<C, F>(
-    connections: &mut HashMap<Arc<[u8]>, C>,
-    cid_routes: &mut HashMap<Arc<[u8]>, Arc<[u8]>>,
-    cid_radix: &mut CidRadix,
-    peer_routes: &mut HashMap<std::net::SocketAddr, Arc<[u8]>>,
-    to_remove: Vec<Arc<[u8]>>,
-    routes_of: F,
-) where
-    F: Fn(&C) -> ConnectionRoutes,
-{
-    for scid in to_remove {
-        if let Some(connection) = connections.remove(&scid) {
-            let routes = routes_of(&connection);
-            purge_connection_routes(
-                cid_routes,
-                cid_radix,
-                peer_routes,
-                &routes.primary_scid,
-                &routes.routing_scids,
-                &routes.peer_address,
-            );
-        }
-    }
-}
-
-/// Deterministically tear down a single stream's in-flight resources.
-///
-/// Drops resources in an order that minimises wasted work:
-/// 1. `body_tx` — signals EOF/cancellation to the upstream H2 task so it
-///    stops reading the request body and exits promptly.
-/// 2. `upstream_result_rx` — dropping the oneshot receiver tells the
-///    upstream task that nobody is listening; `result_tx.send()` in the
-///    spawned task returns `Err` and the task exits.
-/// 3. `response_chunk_rx` — dropping the mpsc receiver causes the
-///    body-pump task's next `chunk_tx.send().await` to return `Err`
-///    and the task exits.
-/// 4. `pending_chunk` — discard any backpressured chunk.
-/// 5. All permits (`global`, `upstream`, `adaptive`, `route_queue`) —
-///    released last so the concurrency slots open up only after we know
-///    no more work is queued.
-///
-/// Returns the phase the stream was in at the time of abort (useful for
-/// logging and metrics at the call site).
-pub(crate) fn abort_stream(req: &mut RequestEnvelope, metrics: &Metrics) -> StreamPhase {
-    let phase = req.phase.clone();
-    if !req.backend_request_finished {
-        if let (Some(pool), Some(index)) = (&req.upstream_pool, req.backend_index)
-            && let Ok(mut guard) = pool.write()
-        {
-            guard.finish_request(
-                index,
-                req.start.elapsed(),
-                req.response_status.or(Some(503)),
-            );
-        }
-        req.backend_request_finished = true;
-    }
-    if req.body_buf_bytes > 0 {
-        metrics.release_request_buffer(req.body_buf_bytes);
-        req.body_buf_bytes = 0;
-    }
-    req.body_buf.clear();
-    req.body_tx = None;
-    req.upstream_result_rx = None;
-    req.response_chunk_rx = None;
-    req.pending_chunk = None;
-    req.global_inflight_permit = None;
-    req.upstream_inflight_permit = None;
-    req.adaptive_admission_permit = None;
-    req.route_queue_permit = None;
-    phase
-}
 
 impl QUICListener {
     pub fn new(config: SpookyConfig) -> Result<Self, ProxyError> {
@@ -3782,389 +3299,6 @@ impl QUICListener {
         ))
     }
 
-    fn pick_alternate_backend(
-        upstream_pool: &Arc<RwLock<UpstreamPool>>,
-        primary_index: usize,
-    ) -> Option<(String, usize)> {
-        let pool = upstream_pool.read().ok()?;
-        for index in pool.pool.healthy_indices_iter() {
-            if index == primary_index {
-                continue;
-            }
-            if let Some(address) = pool.pool.address(index) {
-                return Some((address.to_string(), index));
-            }
-        }
-        None
-    }
-
-    fn log_access(req: &RequestEnvelope, status: u16) {
-        let trace_id = req.trace_id.as_deref().unwrap_or("-");
-        let span_id = req.span_id.as_deref().unwrap_or("-");
-        let latency_ms = req.start.elapsed().as_millis() as u64;
-        if req.routing_transparency_enabled {
-            let reason = if req.routing_transparency_include_reason {
-                req.route_reason.as_deref().unwrap_or("-")
-            } else {
-                "-"
-            };
-            info!(
-                "request_id={} route_upstream={} route_path_len={} route_host_specific={} route_reason={} lb={}",
-                req.request_id,
-                req.upstream_name.as_deref().unwrap_or("-"),
-                req.route_path_len.unwrap_or_default(),
-                req.route_host_specific.unwrap_or(false),
-                reason,
-                req.backend_lb.as_deref().unwrap_or("-")
-            );
-        }
-
-        if let Some(span) = req.trace_span.as_ref() {
-            span.in_scope(|| match req.error_kind.as_ref() {
-                Some(e) => tracing::warn!(
-                    request_id = req.request_id,
-                    trace_id = trace_id,
-                    span_id = span_id,
-                    method = %req.method,
-                    path = %req.path,
-                    status = status,
-                    backend = %req.backend_addr.as_deref().unwrap_or("-"),
-                    upstream = %req.upstream_name.as_deref().unwrap_or("-"),
-                    latency_ms = latency_ms,
-                    retries = req.retry_count,
-                    error = %e,
-                    "request completed with error"
-                ),
-                None => tracing::info!(
-                    request_id = req.request_id,
-                    trace_id = trace_id,
-                    span_id = span_id,
-                    method = %req.method,
-                    path = %req.path,
-                    status = status,
-                    backend = %req.backend_addr.as_deref().unwrap_or("-"),
-                    upstream = %req.upstream_name.as_deref().unwrap_or("-"),
-                    latency_ms = latency_ms,
-                    retries = req.retry_count,
-                    "request completed"
-                ),
-            });
-        }
-
-        match req.error_kind {
-            Some(e) => info!(
-                "request_id={} trace_id={} span_id={} method={} path={} status={} backend={} upstream={} latency_ms={} retries={} error={}",
-                req.request_id,
-                trace_id,
-                span_id,
-                req.method,
-                req.path,
-                status,
-                req.backend_addr.as_deref().unwrap_or("-"),
-                req.upstream_name.as_deref().unwrap_or("-"),
-                latency_ms,
-                req.retry_count,
-                e,
-            ),
-            None => info!(
-                "request_id={} trace_id={} span_id={} method={} path={} status={} backend={} upstream={} latency_ms={} retries={}",
-                req.request_id,
-                trace_id,
-                span_id,
-                req.method,
-                req.path,
-                status,
-                req.backend_addr.as_deref().unwrap_or("-"),
-                req.upstream_name.as_deref().unwrap_or("-"),
-                latency_ms,
-                req.retry_count,
-            ),
-        }
-    }
-
-    /// Handle an already-resolved `ForwardResult`, applying health transitions
-    /// and sending the H3 response.
-    #[allow(clippy::too_many_arguments)]
-    fn handle_forward_result(
-        h3: &mut quiche::h3::Connection,
-        quic: &mut quiche::Connection,
-        stream_id: u64,
-        req: &RequestEnvelope,
-        result: ForwardResult,
-        upstream_pools: &HashMap<String, Arc<RwLock<UpstreamPool>>>,
-        routing_index: &RouteIndex,
-        metrics: &Metrics,
-        overload_retry_after_seconds: u32,
-    ) -> Result<(), quiche::h3::Error> {
-        let start = req.start;
-        let route_label = req.upstream_name.as_deref().unwrap_or("unrouted");
-
-        // If routing failed at Headers time, return an appropriate error now.
-        let (backend_addr, backend_index) = match (&req.backend_addr, req.backend_index) {
-            (Some(a), Some(i)) => (a.as_str(), i),
-            _ => {
-                metrics.inc_failure();
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::Failure);
-                return Self::send_simple_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    if req.method.is_empty() || req.path.is_empty() {
-                        http::StatusCode::BAD_REQUEST
-                    } else {
-                        http::StatusCode::SERVICE_UNAVAILABLE
-                    },
-                    b"no upstream available\n",
-                );
-            }
-        };
-
-        // Re-acquire the upstream pool for health marking.
-        let upstream_name = routing_index.lookup(&req.path, req.authority.as_deref());
-        let upstream_pool = req
-            .upstream_pool
-            .as_ref()
-            .cloned()
-            .or_else(|| upstream_name.and_then(|n| upstream_pools.get(n)).cloned());
-
-        match result {
-            Ok(_) => {
-                error!("Unexpected successful forward result in error handler path");
-                metrics.inc_failure();
-                metrics.inc_backend_error();
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::BackendError);
-                Self::send_simple_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    http::StatusCode::BAD_GATEWAY,
-                    b"unexpected upstream state\n",
-                )
-            }
-            Err(ProxyError::Bridge(err)) => {
-                error!("Bridge error: {:?}", err);
-                metrics.inc_failure();
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::Failure);
-                Self::log_access(req, 400);
-                Self::send_simple_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    http::StatusCode::BAD_REQUEST,
-                    b"invalid request\n",
-                )
-            }
-            Err(ProxyError::Pool(PoolError::BackendOverloaded(reason))) => {
-                metrics.inc_failure();
-                if reason.contains("unknown-length response prebuffer limit exceeded") {
-                    metrics.inc_response_prebuffer_limit_reject();
-                    metrics.inc_overload_shed_reason(OverloadShedReason::ResponsePrebufferCap);
-                } else {
-                    metrics.inc_overload_shed_reason(OverloadShedReason::BackendInflight);
-                }
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::OverloadShed);
-                Self::log_access(req, 503);
-                Self::send_overload_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    b"backend overloaded, retry later\n",
-                    overload_retry_after_seconds,
-                )
-            }
-            Err(ProxyError::Pool(PoolError::CircuitOpen(_))) => {
-                metrics.inc_failure();
-                metrics.inc_circuit_breaker_rejected();
-                metrics.inc_overload_shed_reason(OverloadShedReason::BackendInflight);
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::OverloadShed);
-                Self::log_access(req, 503);
-                Self::send_overload_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    b"backend circuit open, retry later\n",
-                    overload_retry_after_seconds,
-                )
-            }
-            Err(ProxyError::Pool(PoolError::Send(ref send_err))) => {
-                // Log the full inner error — commonly exposes TLS/cert/SNI mismatches
-                // that look like transport failures but are actually local config errors.
-                // Do NOT mark the backend unhealthy: the backend may be fully operational;
-                // the connection failure is a proxy-side config issue.
-                error!(
-                    "Upstream send failed for {} (possible TLS/cert/SNI misconfiguration): {}",
-                    backend_addr, send_err
-                );
-                metrics.inc_health_failure(HealthFailureReason::Tls);
-                metrics.inc_failure();
-                metrics.inc_backend_error();
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::BackendError);
-                Self::log_access(req, 502);
-                Self::send_simple_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    http::StatusCode::BAD_GATEWAY,
-                    b"upstream error\n",
-                )
-            }
-            Err(ProxyError::Transport(_))
-            | Err(ProxyError::Pool(PoolError::InflightLimiterClosed))
-            | Err(ProxyError::Pool(PoolError::UnknownBackend(_))) => {
-                error!("Upstream transport error");
-                metrics.inc_health_failure(HealthFailureReason::Transport);
-                if let Some(pool) = &upstream_pool
-                    && let Some(t) = pool.write().ok().and_then(|mut p| {
-                        p.pool
-                            .mark_request_failure(backend_index, HealthFailureReason::Transport)
-                    })
-                {
-                    Self::log_health_transition(backend_addr, t);
-                }
-                metrics.inc_failure();
-                metrics.inc_backend_error();
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::BackendError);
-                Self::log_access(req, 502);
-                Self::send_simple_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    http::StatusCode::BAD_GATEWAY,
-                    b"upstream error\n",
-                )
-            }
-            Err(ProxyError::Protocol(err)) => {
-                error!("request_id={} Protocol error: {}", req.request_id, err);
-                metrics.inc_failure();
-                metrics.inc_backend_error();
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::BackendError);
-                Self::log_access(req, 502);
-                Self::send_simple_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    http::StatusCode::BAD_GATEWAY,
-                    b"upstream protocol error\n",
-                )
-            }
-            Err(ProxyError::Timeout) => {
-                error!("request_id={} Upstream request timed out", req.request_id);
-                metrics.inc_health_failure(HealthFailureReason::Timeout);
-                if let Some(pool) = &upstream_pool
-                    && let Some(t) = pool.write().ok().and_then(|mut p| {
-                        p.pool
-                            .mark_request_failure(backend_index, HealthFailureReason::Timeout)
-                    })
-                {
-                    Self::log_health_transition(backend_addr, t);
-                }
-                metrics.inc_failure();
-                metrics.inc_timeout();
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::Timeout);
-                Self::log_access(req, 503);
-                Self::send_simple_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    http::StatusCode::SERVICE_UNAVAILABLE,
-                    b"upstream timeout\n",
-                )
-            }
-            Err(ProxyError::Tls(err)) => {
-                error!("TLS error: {}", err);
-                metrics.inc_health_failure(HealthFailureReason::Tls);
-                metrics.inc_failure();
-                metrics.record_route(route_label, start.elapsed(), RouteOutcome::Failure);
-                Self::log_access(req, 500);
-                Self::send_simple_response(
-                    h3,
-                    quic,
-                    stream_id,
-                    http::StatusCode::INTERNAL_SERVER_ERROR,
-                    b"internal server error\n",
-                )
-            }
-        }
-    }
-
-    fn send_simple_response(
-        h3: &mut quiche::h3::Connection,
-        quic: &mut quiche::Connection,
-        stream_id: u64,
-        status: http::StatusCode,
-        body: &[u8],
-    ) -> Result<(), quiche::h3::Error> {
-        let resp_headers = vec![
-            quiche::h3::Header::new(b":status", status.as_str().as_bytes()),
-            quiche::h3::Header::new(b"content-type", b"text/plain"),
-            quiche::h3::Header::new(b"content-length", body.len().to_string().as_bytes()),
-        ];
-
-        h3.send_response(quic, stream_id, &resp_headers, false)?;
-        h3.send_body(quic, stream_id, body, true)?;
-        Ok(())
-    }
-
-    fn send_overload_response(
-        h3: &mut quiche::h3::Connection,
-        quic: &mut quiche::Connection,
-        stream_id: u64,
-        body: &[u8],
-        retry_after_seconds: u32,
-    ) -> Result<(), quiche::h3::Error> {
-        let retry_after = retry_after_seconds.max(1).to_string();
-        let resp_headers = vec![
-            quiche::h3::Header::new(
-                b":status",
-                http::StatusCode::SERVICE_UNAVAILABLE.as_str().as_bytes(),
-            ),
-            quiche::h3::Header::new(b"content-type", b"text/plain"),
-            quiche::h3::Header::new(b"retry-after", retry_after.as_bytes()),
-            quiche::h3::Header::new(b"content-length", body.len().to_string().as_bytes()),
-        ];
-
-        h3.send_response(quic, stream_id, &resp_headers, false)?;
-        h3.send_body(quic, stream_id, body, true)?;
-        Ok(())
-    }
-
-    fn flush_send(socket: &UdpSocket, send_buf: &mut [u8], connection: &mut QuicConnection) {
-        let mut packet_count = 0;
-
-        loop {
-            match connection.quic.send(send_buf) {
-                Ok((write, send_info)) => {
-                    packet_count += 1;
-                    debug!("Sending {} bytes to {}", write, send_info.to);
-                    if let Err(e) = socket.send_to(&send_buf[..write], send_info.to) {
-                        error!("Failed to send UDP packet: {:?}", e);
-                        break;
-                    }
-                }
-                Err(quiche::Error::Done) => break,
-                Err(e) => {
-                    error!("QUIC send failed: {:?}", e);
-                    break;
-                }
-            }
-        }
-
-        if packet_count > 0 {
-            debug!("Sent {} packets", packet_count);
-        }
-    }
-
-    fn log_health_transition(addr: &str, transition: HealthTransition) {
-        match transition {
-            HealthTransition::BecameHealthy => {
-                info!("Backend {} became healthy", addr);
-            }
-            HealthTransition::BecameUnhealthy => {
-                error!("Backend {} became unhealthy", addr);
-            }
-        }
-    }
-
     fn spawn_metrics_endpoint(
         config: &SpookyConfig,
         metrics: Arc<Metrics>,
@@ -4747,638 +3881,6 @@ impl QUICListener {
             Err(_) => Response::new(Full::new(Bytes::from_static(b"failed to render metrics\n"))),
         }
     }
-
-    fn spawn_control_api_endpoint(
-        config: &SpookyConfig,
-        shared_state: &SharedRuntimeState,
-        worker_count: usize,
-    ) -> Result<(), ProxyError> {
-        let endpoint = &config.observability.control_api;
-        if !endpoint.enabled {
-            return Ok(());
-        }
-        let required = endpoint.required;
-
-        let bind = format!("{}:{}", endpoint.address, endpoint.port);
-        let max_connections = endpoint.max_connections.max(1);
-        let connection_timeout = Duration::from_millis(endpoint.connection_timeout_ms.max(1));
-        let paths = ControlApiPaths {
-            health_path: endpoint.health_path.clone(),
-            ready_path: endpoint.ready_path.clone(),
-            runtime_path: endpoint.runtime_path.clone(),
-            restart_path: endpoint.restart_path.clone(),
-        };
-        let state = ControlApiState {
-            metrics: Arc::clone(&shared_state.metrics),
-            resilience: Arc::clone(&shared_state.resilience),
-            watchdog: Arc::clone(&shared_state.watchdog),
-            upstream_pools: shared_state.upstream_pools.clone(),
-            expected_workers: worker_count.max(1),
-            started_at: Instant::now(),
-            auth_token: endpoint.auth_token.clone(),
-        };
-
-        let handle = match runtime_handle() {
-            Some(handle) => handle,
-            None => {
-                let msg = "control API disabled (no Tokio runtime available)".to_string();
-                if required {
-                    return Err(ProxyError::Transport(msg));
-                }
-                error!("{}", msg);
-                return Ok(());
-            }
-        };
-
-        let std_listener = match std::net::TcpListener::bind(&bind) {
-            Ok(listener) => listener,
-            Err(err) => {
-                let msg = format!("failed to bind control API endpoint {bind}: {err}");
-                if required {
-                    return Err(ProxyError::Transport(msg));
-                }
-                error!("{}", msg);
-                return Ok(());
-            }
-        };
-        if let Err(err) = std_listener.set_nonblocking(true) {
-            let msg = format!(
-                "failed to set control API endpoint listener nonblocking ({}): {}",
-                bind, err
-            );
-            if required {
-                return Err(ProxyError::Transport(msg));
-            }
-            error!("{}", msg);
-            return Ok(());
-        }
-        let listener = match tokio::net::TcpListener::from_std(std_listener) {
-            Ok(listener) => listener,
-            Err(err) => {
-                let msg = format!(
-                    "failed to register control API endpoint listener {}: {}",
-                    bind, err
-                );
-                if required {
-                    return Err(ProxyError::Transport(msg));
-                }
-                error!("{}", msg);
-                return Ok(());
-            }
-        };
-
-        spawn_supervised_async_task(
-            &handle,
-            "control-api-endpoint",
-            Some(Arc::clone(&shared_state.metrics)),
-            async move {
-                info!(
-                    "Control API endpoint listening on http://{}{} (ready={}, runtime={}, max_connections={}, connection_timeout_ms={})",
-                    bind,
-                    paths.health_path,
-                    paths.ready_path,
-                    paths.runtime_path,
-                    max_connections,
-                    connection_timeout.as_millis()
-                );
-                let connection_limiter = Arc::new(Semaphore::new(max_connections));
-
-                loop {
-                    let (stream, _peer) = match listener.accept().await {
-                        Ok(v) => v,
-                        Err(err) => {
-                            error!("Control API endpoint accept failed: {}", err);
-                            continue;
-                        }
-                    };
-                    let permit = match Arc::clone(&connection_limiter).try_acquire_owned() {
-                        Ok(permit) => permit,
-                        Err(_) => {
-                            continue;
-                        }
-                    };
-
-                    let io = TokioIo::new(stream);
-                    let paths = paths.clone();
-                    let state = state.clone();
-                    let timeout = connection_timeout;
-
-                    tokio::spawn(async move {
-                        let _permit = permit;
-                        let service = service_fn(move |req: Request<Incoming>| {
-                            let paths = paths.clone();
-                            let state = state.clone();
-                            async move {
-                                Ok::<_, hyper::Error>(Self::handle_control_api_request(
-                                    req, &paths, &state,
-                                ))
-                            }
-                        });
-
-                        let serve = http1::Builder::new().serve_connection(io, service);
-                        match tokio::time::timeout(timeout, serve).await {
-                            Ok(Ok(())) => {}
-                            Ok(Err(err)) => {
-                                error!("Control API endpoint connection failed: {}", err);
-                            }
-                            Err(_) => {
-                                debug!("Control API endpoint connection timed out");
-                            }
-                        }
-                    });
-                }
-            },
-        );
-        Ok(())
-    }
-
-    fn json_response(status: StatusCode, value: serde_json::Value) -> Response<Full<Bytes>> {
-        match Response::builder()
-            .status(status)
-            .header("content-type", "application/json")
-            .body(Full::new(Bytes::from(value.to_string())))
-        {
-            Ok(resp) => resp,
-            Err(_) => Response::new(Full::new(Bytes::from_static(b"{\"error\":\"response\"}"))),
-        }
-    }
-
-    fn handle_control_api_request(
-        req: Request<Incoming>,
-        paths: &ControlApiPaths,
-        state: &ControlApiState,
-    ) -> Response<Full<Bytes>> {
-        let path = req.uri().path();
-
-        if req.method() == http::Method::GET && path == paths.health_path {
-            let response = json!({
-                "status": "ok",
-                "uptime_ms": state.started_at.elapsed().as_millis() as u64,
-                "watchdog": {
-                    "enabled": state.watchdog.enabled(),
-                    "degraded": state.watchdog.is_degraded(),
-                    "restart_requested": state.watchdog.restart_requested(),
-                },
-            });
-            return Self::json_response(StatusCode::OK, response);
-        }
-
-        if req.method() == http::Method::GET && path == paths.ready_path {
-            let (healthy_backends, total_backends) = state.snapshot_backend_health();
-            let restart_requested = state.watchdog.restart_requested();
-            let ready = !restart_requested && (total_backends == 0 || healthy_backends > 0);
-            let response = json!({
-                "ready": ready,
-                "healthy_backends": healthy_backends,
-                "total_backends": total_backends,
-                "restart_requested": restart_requested,
-            });
-            return Self::json_response(
-                if ready {
-                    StatusCode::OK
-                } else {
-                    StatusCode::SERVICE_UNAVAILABLE
-                },
-                response,
-            );
-        }
-
-        if req.method() == http::Method::GET && path == paths.runtime_path {
-            if !Self::control_api_is_authorized(&req, state) {
-                return Self::json_response(
-                    StatusCode::UNAUTHORIZED,
-                    json!({
-                        "error": "unauthorized",
-                    }),
-                );
-            }
-            let (healthy_backends, total_backends) = state.snapshot_backend_health();
-            let response = json!({
-                "uptime_ms": state.started_at.elapsed().as_millis() as u64,
-                "workers": {
-                    "expected": state.expected_workers,
-                },
-                "watchdog": {
-                    "enabled": state.watchdog.enabled(),
-                    "degraded": state.watchdog.is_degraded(),
-                    "restart_requested": state.watchdog.restart_requested(),
-                    "restart_reason": state.watchdog.restart_reason(),
-                    "restart_requested_at_ms": state.watchdog.restart_requested_at_ms(),
-                },
-                "adaptive_admission": {
-                    "enabled": state.resilience.adaptive_admission.enabled(),
-                    "current_limit": state.resilience.adaptive_admission.current_limit(),
-                    "inflight_percent": state.resilience.adaptive_admission.inflight_percent(),
-                },
-                "backends": {
-                    "healthy": healthy_backends,
-                    "total": total_backends,
-                },
-                "metrics": {
-                    "requests_total": state.metrics.requests_total.load(Ordering::Relaxed),
-                    "requests_success": state.metrics.requests_success.load(Ordering::Relaxed),
-                    "requests_failure": state.metrics.requests_failure.load(Ordering::Relaxed),
-                    "active_connections": state.metrics.active_connections.load(Ordering::Relaxed),
-                    "backend_timeouts": state.metrics.backend_timeouts.load(Ordering::Relaxed),
-                    "backend_errors": state.metrics.backend_errors.load(Ordering::Relaxed),
-                },
-                "extension_model": {
-                    "status": "non_goal",
-                    "details": "No plugin/middleware ABI is exposed in-process today; extension support remains a deliberate non-goal until a safe isolation model is designed.",
-                },
-            });
-            return Self::json_response(StatusCode::OK, response);
-        }
-
-        if req.method() == http::Method::POST && path == paths.restart_path {
-            if !Self::control_api_is_authorized(&req, state) {
-                return Self::json_response(
-                    StatusCode::UNAUTHORIZED,
-                    json!({
-                        "accepted": false,
-                        "error": "unauthorized",
-                    }),
-                );
-            }
-            if !state.watchdog.enabled() {
-                return Self::json_response(
-                    StatusCode::SERVICE_UNAVAILABLE,
-                    json!({
-                        "accepted": false,
-                        "error": "watchdog disabled",
-                    }),
-                );
-            }
-
-            let accepted = state.watchdog.request_restart("admin_runtime_api");
-            return Self::json_response(
-                if accepted {
-                    StatusCode::ACCEPTED
-                } else {
-                    StatusCode::CONFLICT
-                },
-                json!({
-                    "accepted": accepted,
-                    "restart_requested": state.watchdog.restart_requested(),
-                    "reason": if accepted { "admin_runtime_api" } else { "restart pending or cooldown active" },
-                }),
-            );
-        }
-
-        match Response::builder()
-            .status(StatusCode::NOT_FOUND)
-            .body(Full::new(Bytes::from_static(b"not found\n")))
-        {
-            Ok(resp) => resp,
-            Err(_) => Response::new(Full::new(Bytes::from_static(b"not found\n"))),
-        }
-    }
-
-    fn control_api_is_authorized(req: &Request<Incoming>, state: &ControlApiState) -> bool {
-        let Some(token) = state.auth_token.as_ref() else {
-            return true;
-        };
-        let Some(header) = req.headers().get(http::header::AUTHORIZATION) else {
-            return false;
-        };
-        let Ok(raw) = header.to_str() else {
-            return false;
-        };
-        let Some(provided) = raw.strip_prefix("Bearer ") else {
-            return false;
-        };
-        provided == token
-    }
-
-    fn spawn_watchdog(
-        config: &SpookyConfig,
-        metrics: Arc<Metrics>,
-        resilience: Arc<RuntimeResilience>,
-        watchdog: Arc<WatchdogCoordinator>,
-    ) {
-        let watchdog_config = WatchdogRuntimeConfig::from(&config.resilience.watchdog);
-        if !watchdog_config.enabled || !watchdog.enabled() {
-            return;
-        }
-
-        let handle = match runtime_handle() {
-            Some(handle) => handle,
-            None => {
-                error!("Watchdog disabled: no Tokio runtime available");
-                return;
-            }
-        };
-
-        spawn_supervised_async_task(
-            &handle,
-            "watchdog",
-            Some(Arc::clone(&metrics)),
-            async move {
-                info!(
-                    "Watchdog enabled: check_interval_ms={} poll_stall_timeout_ms={} timeout_error_rate_percent={} overload_inflight_percent={} unhealthy_windows={} drain_grace_ms={} restart_cooldown_ms={}",
-                    watchdog_config.check_interval_ms,
-                    watchdog_config.poll_stall_timeout_ms,
-                    watchdog_config.timeout_error_rate_percent,
-                    watchdog_config.overload_inflight_percent,
-                    watchdog_config.unhealthy_consecutive_windows,
-                    watchdog_config.drain_grace_ms,
-                    watchdog_config.restart_cooldown_ms,
-                );
-
-                let mut interval =
-                    tokio::time::interval(Duration::from_millis(watchdog_config.check_interval_ms));
-                interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-                let restart_program = watchdog_config
-                    .restart_command
-                    .first()
-                    .map(|value| value.trim().to_string())
-                    .filter(|value| !value.is_empty());
-                let has_restart_command = restart_program.is_some();
-                if watchdog_config
-                    .restart_hook
-                    .as_deref()
-                    .map(str::trim)
-                    .is_some_and(|value| !value.is_empty())
-                {
-                    warn!(
-                        "Watchdog restart_hook is deprecated and ignored; configure resilience.watchdog.restart_command instead"
-                    );
-                }
-
-                let mut previous_requests = metrics.requests_total.load(Ordering::Relaxed);
-                let mut previous_timeouts = metrics.backend_timeouts.load(Ordering::Relaxed);
-                let mut degraded_windows = 0u32;
-
-                loop {
-                    interval.tick().await;
-                    let now = now_millis();
-                    let stalled = now.saturating_sub(watchdog.last_poll_progress_ms())
-                        > watchdog_config.poll_stall_timeout_ms;
-
-                    let current_requests = metrics.requests_total.load(Ordering::Relaxed);
-                    let current_timeouts = metrics.backend_timeouts.load(Ordering::Relaxed);
-                    let request_delta = current_requests.saturating_sub(previous_requests);
-                    let timeout_delta = current_timeouts.saturating_sub(previous_timeouts);
-                    previous_requests = current_requests;
-                    previous_timeouts = current_timeouts;
-
-                    let timeout_rate_percent = timeout_delta
-                        .saturating_mul(100)
-                        .checked_div(request_delta)
-                        .unwrap_or(0);
-
-                    let timeout_pressure = request_delta >= watchdog_config.min_requests_per_window
-                        && timeout_rate_percent
-                            >= watchdog_config.timeout_error_rate_percent as u64;
-                    let overload_pressure = resilience.adaptive_admission.inflight_percent()
-                        >= watchdog_config.overload_inflight_percent;
-
-                    if stalled || timeout_pressure || overload_pressure {
-                        degraded_windows = degraded_windows.saturating_add(1);
-                        watchdog.set_degraded(true);
-                        metrics.inc_watchdog_degraded_window();
-                    } else {
-                        degraded_windows = 0;
-                        watchdog.set_degraded(false);
-                    }
-
-                    if degraded_windows >= watchdog_config.unhealthy_consecutive_windows {
-                        if !has_restart_command {
-                            warn!(
-                                "Watchdog detected unhealthy runtime state, but restart_command is not configured"
-                            );
-                            degraded_windows = 0;
-                            continue;
-                        }
-                        let mut reasons = Vec::new();
-                        if stalled {
-                            reasons.push("poll_stall");
-                        }
-                        if timeout_pressure {
-                            reasons.push("timeout_spike");
-                        }
-                        if overload_pressure {
-                            reasons.push("inflight_overload");
-                        }
-                        let reason = reasons.join("+");
-                        if watchdog.request_restart(&reason) {
-                            metrics.inc_watchdog_restart_request();
-                            warn!("Watchdog requested safe restart: {}", reason);
-                        }
-                        degraded_windows = 0;
-                    }
-
-                    if !watchdog.restart_requested() {
-                        continue;
-                    }
-
-                    let requested_at = watchdog.restart_requested_at_ms();
-                    let grace_elapsed = requested_at != 0
-                        && now.saturating_sub(requested_at) >= watchdog_config.drain_grace_ms;
-                    if !watchdog.workers_drained() && !grace_elapsed {
-                        continue;
-                    }
-
-                    let restart_reason = watchdog.restart_reason();
-                    if watchdog.workers_drained() {
-                        info!(
-                            "Watchdog safe restart condition reached (all workers drained): {}",
-                            restart_reason
-                        );
-                    } else {
-                        warn!(
-                            "Watchdog restart drain grace elapsed; executing hook without full drain: {}",
-                            restart_reason
-                        );
-                    }
-
-                    let program = restart_program.as_deref().unwrap_or_default();
-                    let args: Vec<&str> = watchdog_config
-                        .restart_command
-                        .iter()
-                        .skip(1)
-                        .map(String::as_str)
-                        .collect();
-                    let status = tokio::process::Command::new(program)
-                        .args(args)
-                        .env("SPOOKY_WATCHDOG_REASON", &restart_reason)
-                        .status()
-                        .await;
-                    match status {
-                        Ok(status) => {
-                            info!(
-                                "Watchdog restart hook exited with status {}",
-                                status
-                                    .code()
-                                    .map(|code| code.to_string())
-                                    .unwrap_or_else(|| "signal".to_string())
-                            );
-                        }
-                        Err(err) => {
-                            error!("Watchdog restart hook execution failed: {}", err);
-                        }
-                    }
-                    metrics.inc_watchdog_restart_hook();
-
-                    watchdog.complete_restart_cycle();
-                }
-            },
-        );
-    }
-
-    fn spawn_health_checks(
-        upstream_pools: HashMap<String, Arc<RwLock<UpstreamPool>>>,
-        health_client: Arc<H2Client>,
-        metrics: Arc<Metrics>,
-    ) {
-        struct HealthCheckJob {
-            upstream_pool: Arc<RwLock<UpstreamPool>>,
-            index: usize,
-            address: String,
-            health_uri: String,
-            timeout: Duration,
-            base_interval_ms: u64,
-            consecutive_failures: u32,
-            next_due_at: Instant,
-        }
-
-        let mut grouped_jobs: HashMap<u64, Vec<HealthCheckJob>> = HashMap::new();
-        for (_upstream_name, upstream_pool) in upstream_pools.iter() {
-            let pool = match upstream_pool.read() {
-                Ok(pool) => pool,
-                Err(_) => continue,
-            };
-            for index in pool.pool.all_indices() {
-                let (address, health) =
-                    match (pool.pool.address(index), pool.pool.health_check(index)) {
-                        (Some(address), Some(health)) => (address.to_string(), health),
-                        _ => continue,
-                    };
-                let path: &str = if health.path.is_empty() {
-                    "/"
-                } else {
-                    &health.path
-                };
-                let endpoint = match BackendEndpoint::parse(&address) {
-                    Ok(endpoint) => endpoint,
-                    Err(err) => {
-                        error!(
-                            "disabling health checks for backend '{}' due to invalid endpoint: {}",
-                            address, err
-                        );
-                        continue;
-                    }
-                };
-                let base_interval_ms = health.interval.max(1);
-                let initial_jitter_ms = if base_interval_ms > 1 {
-                    crate::stable_hash64(address.as_bytes()) % base_interval_ms
-                } else {
-                    0
-                };
-                let next_due_at = Instant::now() + Duration::from_millis(initial_jitter_ms);
-                let job = HealthCheckJob {
-                    upstream_pool: Arc::clone(upstream_pool),
-                    index,
-                    address,
-                    health_uri: endpoint.uri_for_path(path),
-                    timeout: Duration::from_millis(health.timeout_ms.max(1)),
-                    base_interval_ms,
-                    consecutive_failures: 0,
-                    next_due_at,
-                };
-                grouped_jobs.entry(base_interval_ms).or_default().push(job);
-            }
-        }
-
-        if grouped_jobs.is_empty() {
-            return;
-        }
-
-        let handle = match runtime_handle() {
-            Some(handle) => handle,
-            None => {
-                error!("Health checks disabled: no Tokio runtime available");
-                return;
-            }
-        };
-
-        for (base_interval_ms, mut jobs) in grouped_jobs {
-            let health_client = Arc::clone(&health_client);
-            let task_metrics = Arc::clone(&metrics);
-            let handle = handle.clone();
-            let supervise_metrics = Arc::clone(&task_metrics);
-            spawn_supervised_async_task(
-                &handle,
-                "health-check-group",
-                Some(supervise_metrics),
-                async move {
-                    let scheduler_tick_ms = (base_interval_ms / 4).clamp(20, base_interval_ms);
-                    let mut ticker =
-                        tokio::time::interval(Duration::from_millis(scheduler_tick_ms.max(1)));
-                    ticker.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-
-                    loop {
-                        ticker.tick().await;
-                        let now = Instant::now();
-                        for job in jobs.iter_mut() {
-                            if now < job.next_due_at {
-                                continue;
-                            }
-
-                            let request = match http::Request::builder()
-                                .method("GET")
-                                .uri(&job.health_uri)
-                                .body(BoxBody::new(Full::new(Bytes::new())))
-                            {
-                                Ok(req) => req,
-                                Err(_) => continue,
-                            };
-
-                            let result =
-                                tokio::time::timeout(job.timeout, health_client.send(request))
-                                    .await;
-
-                            let outcome = match result {
-                                Ok(Ok(response)) => {
-                                    classify_active_health_check_response(response.status())
-                                }
-                                _ => HealthClassification::Failure,
-                            };
-
-                            let transition = match job.upstream_pool.write() {
-                                Ok(mut pool) => match outcome {
-                                    HealthClassification::Success => {
-                                        task_metrics.inc_health_check_success();
-                                        job.consecutive_failures = 0;
-                                        pool.pool.mark_success(job.index)
-                                    }
-                                    HealthClassification::Failure => {
-                                        task_metrics.inc_health_check_failure();
-                                        job.consecutive_failures =
-                                            job.consecutive_failures.saturating_add(1);
-                                        pool.pool.mark_failure(job.index)
-                                    }
-                                    HealthClassification::Neutral => {
-                                        job.consecutive_failures = 0;
-                                        None
-                                    }
-                                },
-                                Err(_) => None,
-                            };
-
-                            let backoff_multiplier = 1u64 << job.consecutive_failures.min(2);
-                            let delay_ms = job.base_interval_ms.saturating_mul(backoff_multiplier);
-                            job.next_due_at = Instant::now() + Duration::from_millis(delay_ms);
-
-                            if let Some(transition) = transition {
-                                Self::log_health_transition(&job.address, transition);
-                            }
-                        }
-                    }
-                },
-            );
-        }
-    }
 }
 
 fn classify_retry_reason(err: &ProxyError) -> RetryReason {
@@ -5388,10 +3890,6 @@ fn classify_retry_reason(err: &ProxyError) -> RetryReason {
         ProxyError::Pool(_) => RetryReason::BackendPool,
         _ => RetryReason::BackendTransport,
     }
-}
-
-fn classify_active_health_check_response(status: StatusCode) -> HealthClassification {
-    outcome_from_status(status)
 }
 
 pub fn configure_async_runtime(worker_threads: usize) {
